+++ R/spautor.R
@@ -11,14 +11,15 @@
 #'   the variables in \code{fixed}, \code{random}, and \code{partition_factor},
 #'   as well as potentially geographical information.
 #' @param spcov_type The spatial covariance type. Available options include
-#'   \code{"car"} and \code{"sar"}. Parameterizations of each spatial covariance type are
+#'   \code{"car"}, \code{"sar"}, and \code{"none"}. Parameterizations of each spatial covariance type are
 #'   available in Details. When \code{spcov_type} is specified, relevant spatial
 #'   covariance parameters are assumed unknown, requiring estimation.
 #'   \code{spcov_type} is not required (and is
 #'   ignored) if \code{spcov_initial} is provided.  Multiple values can be
 #'   provided in a character vector. Then \code{spautor()} is called iteratively
 #'   for each element and a list is returned for each model fit.
-#'   The default for \code{spcov_type} is \code{"car"}.
+#'   The default for \code{spcov_type} is \code{"car"}. When \code{spcov_type}
+#'   is \code{"none"}, [splm()] is called.
 #' @param spcov_initial An object from [spcov_initial()] specifying initial and/or
 #'   known values for the spatial covariance parameters.
 #'   Not required if \code{spcov_type} is provided. Multiple [spcov_initial()]
@@ -91,6 +92,7 @@
 #'      symmetry condition matrix \eqn{M}
 #'     \item sar: \eqn{[(I - range * W)(I - range * W)^T]^{-1}},
 #'      weights matrix \eqn{W}, \eqn{^T} indicates matrix transpose
+#'     \item none: \eqn{0}
 #'   }
 #'   If there are observations with no neighbors, they are given a unique variance
 #'   parameter called \code{extra}, which must be non-negative.
@@ -185,6 +187,19 @@ spautor <- function(formula, data, spcov_type, spcov_initial, estmethod = "reml"
     names(spautor_out) <- spcov_type
     new_spautor_out <- structure(spautor_out, class = "spautor_list")
     return(new_spautor_out)
+  }
+
+
+
+  # call splm if spcov_type is none (works on an individual element of spcov_type or initial)
+  if ((!missing(spcov_type) && spcov_type %in% c("none", "ie")) || (!missing(spcov_initial) && inherits(spcov_initial, c("none", "ie")))) {
+    call_list <- as.list(match.call())[-1]
+    # remove spautor specific arguments
+    args_remove <- c("W", "row_st", "range_positive", "cutoff")
+    call_list <- call_list[!names(call_list) %in% args_remove]
+    penv <- parent.frame()
+    splm_out <- do.call("splm", call_list, envir = penv)
+    return(splm_out)
   }
 
   # replace initial values with appropriate NA's