+++ NEWS.md
@@ -1,3 +1,12 @@
+# spmodel 0.10.0
+
+## Minor updates
+
+* Added a robust semivariogram option to `esv()`; see the `robust`argument to `esv()` ([#28](https://github.com/USEPA/spmodel/issues/28)).
+* Added `"none"` and `"ie"` spatial covariance types (via `spcov_type` or `spcov_initial`) to `spautor()`, `spgautor()`, and `spautorRF()` ([#27](https://github.com/USEPA/spmodel/issues/27)).
+* Fixed the aspect ratio at one for level curve plots of fitted model objects (anisotropic and isotropic).
+* Changed the title for level curve plots of fitted model objects from "Anisotropic level curve" to "Isotropic level curve" when the model is isotropic.
+
 # spmodel 0.9.0
 
 ## Major Updates